fix(types): consolidate IPC type drifts to canonical sources
- Move BranchInfo and CreateWorktreeOptions to shared/types/git.ts and
  re-export from workspace-host.ts and ipc/worktree.ts; drop duplicates
  in electron/services/GitService.ts (Issue #7280, drifts 1-2)
- Narrow terminal-diagnostic-info event payload from 'any' to
  TerminalInfoPayload | null in shared/types/pty-host.ts (drift 3)
- Re-export AgentState and AgentStateChangeTrigger from canonical
  shared/types/agent.ts in shared/types/ipc/agent.ts (drift 4)
- Narrow RecoveryAction.actionId from string to ActionId open union
  in shared/types/ipc/errors.ts (drift 5)
- Replace inline 4-state union for voice-input:status with
  VoiceInputStatus (5 states, includes 'finishing') and narrow
  terminal:get-by-state args to AgentState in shared/types/ipc/maps.ts
  (drifts 6-7)
- Wrap FileReadErrorCode with _AssertSubset<AppErrorCode, ...> guard
  so a removed AppErrorCode member fails typecheck loudly rather than
  silently shrinking (drift 8)
