+++ shared/types/git.ts
@@ -102,3 +102,24 @@ export interface StagingStatus {
   /** When `repoState === "REBASING"`, the total step count. Null otherwise. */
   rebaseTotalSteps: number | null;
 }
+
+/** Branch information from git */
+export interface BranchInfo {
+  name: string;
+  current: boolean;
+  commit: string;
+  remote?: string;
+}
+
+/** Options for creating a new worktree */
+export interface CreateWorktreeOptions {
+  baseBranch: string;
+  newBranch: string;
+  path: string;
+  fromRemote?: boolean;
+  useExistingBranch?: boolean;
+  /** Opt-in flag to run resource.provision after setup */
+  provisionResource?: boolean;
+  /** Worktree environment mode ("local" or an environment key from resourceEnvironments) */
+  worktreeMode?: string;
+}
